add build_scenario and dag_run_generic scripts
build_scenario.py generates a scenario json file based on the separate files
that already exist, with some capability to generate missing pieces randomly
(WIP)

Commandline arguments were added to generic.cpp for use in the dag_run script,
as well as a few other minor changes & cleanup. It now takes a 'workflowFile'
override in the scenario json to use a different file for the workflow.

json2topo was also added, and topo2json was moved here for use in
build_scenario.py

critial-path-metric was renamed to critial_path_metric, since python requires
names to be a single token for importing, and usage was also updated in the
dag_run scripts